AB#112

+++ Singer.API/Controllers/LegalGuardianUserController.cs
@@ -6,7 +6,7 @@ using Microsoft.AspNetCore.Mvc;
 using Singer.DTOs.Users;
 using Singer.Helpers.Exceptions;
 using Singer.Models.Users;
-using Singer.Services;
+using Singer.Resources;
 using Singer.Services.Interfaces;
 
 namespace Singer.Controllers
@@ -16,10 +16,12 @@ namespace Singer.Controllers
    public class LegalGuardianUserController : DataControllerBase<LegalGuardianUser, LegalGuardianUserDTO, CreateLegalGuardianUserDTO, UpdateLegalGuardianUserDTO>
    {
       private readonly ILegalGuardianUserService _legalGuardianUserService;
+
       public LegalGuardianUserController(ILegalGuardianUserService legalGuardianUserService) : base(legalGuardianUserService)
       {
          _legalGuardianUserService = legalGuardianUserService;
       }
+
       [HttpPut("{id}")]
       [ProducesResponseType(StatusCodes.Status200OK)]
       [ProducesResponseType(StatusCodes.Status404NotFound)]
@@ -27,19 +29,17 @@ namespace Singer.Controllers
       public override async Task<IActionResult> Update(Guid id, [FromBody]UpdateLegalGuardianUserDTO dto)
       {
          if (dto is null)
-         {
-            throw new BadInputException(nameof(dto));
-         }
+            throw new BadInputException("No dto was passed in the body of the request", ErrorMessages.NoDataPassed);
+
+         var model = ModelState;
+         if (!model.IsValid)
+            return BadRequest(model);
 
          if ((dto.CareUsersToAdd?.Count ?? 0) > 0)
-         {
             await _legalGuardianUserService.AddLinkedUsers(id, dto.CareUsersToAdd);
-         }
 
          if ((dto.CareUsersToRemove?.Count ?? 0) > 0)
-         {
             await _legalGuardianUserService.RemoveLinkedUsers(id, dto.CareUsersToRemove);
-         }
 
          var result = await DatabaseService.UpdateAsync(id, dto);
          return Ok(result);
